Refactor import paths in App.js, ItemsList.js, and babel.config.js
