+++ package.json
@@ -1,15 +1,21 @@
 {
+  "name": "activity-and-diet-tracker-app",
+  "version": "1.0.0",
+  "main": "expo/AppEntry.js",
+  "scripts": {
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web"
+  },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.3",
-    "@react-native-community/datetimepicker": "8.0.1",
-    "@react-navigation/bottom-tabs": "^6.6.1",
-    "@react-navigation/native": "^6.1.18",
-    "@react-navigation/native-stack": "^6.11.0",
-    "expo": "^51.0.37",
+    "expo": "~51.0.28",
+    "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
-    "react-native": "0.74.5",
-    "react-native-dropdown-select-list": "^2.0.5",
-    "react-native-safe-area-context": "4.10.5",
-    "react-native-screens": "3.31.1"
-  }
+    "react-native": "0.74.5"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.0"
+  },
+  "private": true
 }
